+++ src/mongo/util/concurrency/ticketholder.cpp
@@ -63,7 +63,7 @@ void updateQueueStatsOnTicketAcquisition(ServiceContext* serviceContext,
 }
 }  // namespace
 
-void TicketHolderWithQueueingStats::resize(OperationContext* opCtx, int newSize) noexcept {
+void TicketHolderWithQueueingStats::resize(OperationContext* opCtx, int32_t newSize) noexcept {
     stdx::lock_guard<Latch> lk(_resizeMutex);
 
     _resize(opCtx, newSize, _outof.load());
@@ -147,7 +147,7 @@ boost::optional<Ticket> TicketHolderWithQueueingStats::waitForTicketUntil(Operat
     }
 }
 
-int TicketHolderWithQueueingStats::getAndResetPeakUsed() {
+int32_t TicketHolderWithQueueingStats::getAndResetPeakUsed() {
     return _peakUsed.swap(used());
 }
 
@@ -170,12 +170,12 @@ void TicketHolderWithQueueingStats::_appendCommonQueueImplStats(BSONObjBuilder& 
 
     b.append("addedToQueue", added);
     b.append("removedFromQueue", removed);
-    b.append("queueLength", std::max(static_cast<int>(added - removed), 0));
+    b.append("queueLength", std::max(added - removed, (int64_t)0));
 
     auto finished = stats.totalFinishedProcessing.loadRelaxed();
     auto started = stats.totalStartedProcessing.loadRelaxed();
     b.append("startedProcessing", started);
-    b.append("processing", std::max(static_cast<int>(started - finished), 0));
+    b.append("processing", std::max(started - finished, (int64_t)0));
     b.append("finishedProcessing", finished);
     b.append("totalTimeProcessingMicros", stats.totalTimeProcessingMicros.loadRelaxed());
     b.append("canceled", stats.totalCanceled.loadRelaxed());
